LayoutTests/editing/style/style-boundary-001.html

<!DOCTYPE html>
<html>
<head>
<style>
.editing { 
    border: 2px solid red; 
    font-size: 24px; 
}
.explanation { 
    border: 2px solid blue; 
    padding: 12px; 
    font-size: 24px; 
    margin-bottom: 24px;
}
.scenario { margin-bottom: 16px;}
.scenario:first-line { font-weight: bold; margin-bottom: 16px;}
.expected-results:first-line { font-weight: bold }
</style>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests bolding a character immediately after a character that is already in bold.');

function editingTest() {
    moveSelectionForwardByCharacterCommand();
    extendSelectionForwardByCharacterCommand();
    Markup.dump('root', 'Before bolding');
    boldCommand();
    Markup.dump('root', 'After bolding');
}

</script>

<title>Editing Test</title>
</head>
<body>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div id="test" class="editing"><b>a</b>b</div>
</div>

<script>
runEditingTestWithCallbackLogging(false);
</script>

</body>
</html>
